Moving all listeners to same project. Now they are triggered again.
