+++ listeners/src/main/java/org/fao/geonet/listener/history/GenericMetadataEventListener.java
@@ -20,7 +20,7 @@
  * Contact: Jeroen Ticheler - FAO - Viale delle Terme di Caracalla 2,
  * Rome - Italy. email: geonetwork@osgeo.org
  */
-package org.fao.geonet.listeners.history;
+package org.fao.geonet.listener.history;
 
 import org.fao.geonet.constants.Geonet;
 import org.fao.geonet.domain.ISODate;
