+++ listeners/src/main/java/org/fao/geonet/listener/history/RecordPrivilegesChangeListener.java
@@ -20,7 +20,7 @@
  * Contact: Jeroen Ticheler - FAO - Viale delle Terme di Caracalla 2,
  * Rome - Italy. email: geonetwork@osgeo.org
  */
-package org.fao.geonet.listeners.history;
+package org.fao.geonet.listener.history;
 
 import org.fao.geonet.domain.StatusValue;
 import org.fao.geonet.events.history.RecordPrivilegesChangeEvent;
